backend/mt5_service/requirements.txt
MetaTrader5>=5.0.37
supabase>=1.0.0
python-dotenv>=1.0.0
